Refactor component 'CreateTodoList' component 'ActionItemAddTodoList'
