+++ react-app/src/js/dashboard/components/ActionBar.js
@@ -1,9 +1,9 @@
 import React from "react";
 
-import CreateTodoList from "./CreateTodoList";
+import ActionItemAddTodoList from "./ActionItemAddTodoList";
 
 export default class ActionBar extends React.Component {
   render() {
-    return <CreateTodoList />;
+    return <ActionItemAddTodoList />;
   }
 }
